@@ -8,31 +8,31 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^8.1.0",
+        "php": "^8.2",
         "guzzlehttp/guzzle": "^7.5",
-        "hammerstone/fast-paginate": "^0.1.11",
+        "hammerstone/fast-paginate": "^0.1.12",
         "larabug/larabug": "^3.0",
         "laravel/fortify": "^1.16",
-        "laravel/framework": "^10.0",
+        "laravel/framework": "^10.4.1",
         "laravel/sanctum": "3.2",
-        "laravel/tinker": "^2.8",
-        "league/flysystem-aws-s3-v3": "^3.12.1",
-        "livewire/livewire": "^2.11",
+        "laravel/tinker": "^2.8.1",
+        "league/flysystem-aws-s3-v3": "^3.12.2",
+        "livewire/livewire": "^2.12.3",
         "lukeraymonddowning/honey": "^0.5.0",
         "torchlight/torchlight-commonmark": "^0.5.5"
     },
     "require-dev": {
-        "barryvdh/laravel-debugbar": "^3.7",
+        "barryvdh/laravel-debugbar": "^3.8.1",
         "fakerphp/faker": "^1.21",
         "laravel/pint": "^0.2.4",
-        "laravel/sail": "^1.18.1",
+        "laravel/sail": "^1.21.2",
         "mockery/mockery": "^1.5.1",
-        "nunomaduro/collision": "^7.0",
+        "nunomaduro/collision": "^7.1.2",
         "nunomaduro/larastan": "^2.5.1",
         "nunomaduro/patrol": "^1.1",
-        "phpunit/phpunit": "^10.0",
-        "spatie/laravel-ignition": "^2.0",
-        "spatie/laravel-ray": "^1.32"
+        "phpunit/phpunit": "^10.0.16",
+        "rector/rector": "^0.15.21",
+        "spatie/laravel-ignition": "^2.0"
     },
     "config": {
         "optimize-autoloader": true,
@@ -42,6 +42,11 @@
     "extra": {
         "laravel": {
             "dont-discover": []
+        },
+        "phpstan": {
+            "includes": [
+                "extension.neon"
+            ]
         }
     },
     "autoload": {
