@@ -14,26 +14,22 @@ class AnimationsServiceProvider extends ServiceProvider
 
     public function boot()
     {
-        Blade::directive('slideUp', function () {
-            return <<< 'ANIMATE'
+        Blade::directive('slideUp', fn () => <<< 'ANIMATE'
                 x-transition:enter="ease-in-out duration-700"
                 x-transition:enter-start="translate-y-full"
                 x-transition:enter-end="translate-y-0"
                 x-transition:leave="ease-in-out duration-300"
                 x-transition:leave-start="translate-y-0"
                 x-transition:leave-end="translate-y-full"
-            ANIMATE;
-        });
+            ANIMATE);
 
-        Blade::directive('showHide', function () {
-            return <<< 'ANIMATE'
+        Blade::directive('showHide', fn () => <<< 'ANIMATE'
                x-transition:enter="ease-out duration-200"
                x-transition:enter-start="opacity-0 scale-95"
                x-transition:enter-end="opacity-100 scale-100"
                x-transition:leave="ease-in duration-75"
                x-transition:leave-start="opacity-100 scale-100"
                x-transition:leave-end="opacity-0 scale-95"
-            ANIMATE;
-        });
+            ANIMATE);
     }
 }
