@@ -8,6 +8,6 @@ trait PasswordValidationRules
 {
     protected function passwordRules()
     {
-        return ['required', 'string', new Password, 'confirmed'];
+        return ['required', 'string', new Password(), 'confirmed'];
     }
 }
