@@ -1,3 +1,6 @@
-<div {{ $attributes }} class="p-4 border-b mb-6 border-slate-200 dark:border-slate-600">
+<div
+    class="p-4 border-b mb-6 border-slate-200 dark:border-slate-600"
+    {{ $attributes }}
+>
     <span class="font-medium text-xl leading-6 text-gray-900 dark:text-slate-300">{{ $slot }}</span>
 </div>
