@@ -1,12 +1,12 @@
-@if($errors->any())
+@if ($errors->any())
     <div>
         <div>
-            {{--Icon here--}}
+            {{-- Icon here --}}
         </div>
         <div>
             <h2>{{ $title ?? $errorCountTitle($errors) }}</h2>
             <ul>
-                @foreach($errors->all() as $error)
+                @foreach ($errors->all() as $error)
                     <li>{{ $error }}</li>
                 @endforeach
             </ul>
