@@ -1,21 +1,26 @@
-{{--Template--}}
-<select {{ $attributes }}
-        autocomplete="off"
-        id="{{ $id }}"
-        multiple
-        name="{{ $name ?? Str::singular($id) }}_id[]"
-        x-data
-        x-init="multiSelect($refs.select)"
-        x-ref="select">
+{{-- Template --}}
+<select
+    id="{{ $id }}"
+    name="{{ $name ?? Str::singular($id) }}_id[]"
+    {{ $attributes }}
+    autocomplete="off"
+    multiple
+    x-data
+    x-init="multiSelect($refs.select)"
+    x-ref="select"
+>
     {{ $slot }}
 </select>
-{{--End Template--}}
+{{-- End Template --}}
 
-{{--Scripts--}}
+{{-- Scripts --}}
 @once
     <script src="https://cdn.jsdelivr.net/npm/tom-select@1.7/dist/js/tom-select.complete.min.js"></script>
 
-    <script async defer>
+    <script
+        async
+        defer
+    >
         const multiSelect = (ref) => {
             const options = {
                 create: true,
@@ -26,10 +31,9 @@
         };
     </script>
 @endonce
-{{--End Scripts--}}
+{{-- End Scripts --}}
 
-{{--Styles--}}
+{{-- Styles --}}
 @push('stylesheets')
-
 @endpush
-{{--End Styles--}}
+{{-- End Styles --}}
