@@ -1,10 +1,12 @@
 @props(['for'])
 
-<textarea {{ $attributes->merge(['id' => $for, 'name' => $for]) }}
-          class="@if ($errors->has($for)) form-pizza-error @else form-pizza @endif w-full sm:w-1/2 mt-2.5">
+<textarea
+    class="@if ($errors->has($for)) form-pizza-error @else form-pizza @endif w-full sm:w-1/2 mt-2.5"
+    {{ $attributes->merge(['id' => $for, 'name' => $for]) }}
+>
     {{ old($for, $slot ?? '') }}
 </textarea>
 
 @error($for)
-<p class="text-red-500 text-sm mt-2.5">{{ $message }}</p>
+    <p class="text-red-500 text-sm mt-2.5">{{ $message }}</p>
 @enderror
