@@ -1,9 +1,14 @@
+@php use App\Enums\PostStatus; @endphp
 <x-fos.layout title="{{ $setTitle }}">
     <x-fos.content>
         <x-auth.navigation></x-auth.navigation>
         <div class="space-y-6">
-            <x-form.form action="{{ $action }}" enctype="multipart/form-data" method="post">
-                @if($mode === 'edit')
+            <x-form.form
+                    action="{{ $action }}"
+                    enctype="multipart/form-data"
+                    method="post"
+            >
+                @if ($mode === 'edit')
                     @method('PATCH')
                 @endif
                 <div class="space-y-10">
@@ -13,7 +18,7 @@
                                 <h3 class="text-xl font-medium leading-6 text-gray-800 dark:text-gray-400">Post Information</h3>
                             </div>
 
-                            {{--Title--}}
+                            {{-- Title --}}
                             <div class="space-y-8 mt-5 md:mt-0 md:col-span-2">
                                 <div class="grid grid-cols-2 gap-6">
                                     <div class="col-span-3 sm:col-span-2">
@@ -23,40 +28,53 @@
                                     </div>
                                 </div>
 
-                                {{--Categories--}}
+                                {{-- Categories --}}
                                 <div>
                                     <x-form.label for="category">
                                         <x-form.select for="category_id">
-                                            <option disabled selected value="0">--- Choose Category ---</option>
-                                            @foreach($categories as $category)
-                                                @if($mode === 'create')
-                                                    <option {{ old('category_id') ? 'selected' : ''  }}
-                                                            value="{{ $category->id }}">{{ $category->name }}</option>
+                                            <option
+                                                    value="0"
+                                                    disabled
+                                                    selected
+                                            >--- Choose Category ---
+                                            </option>
+                                            @foreach ($categories as $category)
+                                                @if ($mode === 'create')
+                                                    <option
+                                                            value="{{ $category->id }}"
+                                                            {{ old('category_id') ? 'selected' : '' }}
+                                                    >{{ $category->name }}</option>
                                                 @endif
 
-                                                @if($mode === 'edit')
-                                                    <option {{ $post->category->id === $category->id ? 'selected' : ''  }}
-                                                            value="{{ $category->id }}">{{ $category->name }}</option>
+                                                @if ($mode === 'edit')
+                                                    <option
+                                                            value="{{ $category->id }}"
+                                                            {{ $post->category->id === $category->id ? 'selected' : '' }}
+                                                    >{{ $category->name }}</option>
                                                 @endif
                                             @endforeach
                                         </x-form.select>
                                     </x-form.label>
                                 </div>
 
-                                {{--Excerpt--}}
+                                {{-- Excerpt --}}
                                 <div>
                                     <x-form.label for="excerpt">
-                                        <x-form.input for="excerpt" type="text" value="{{ old('excerpt', $post->excerpt) }}" />
+                                        <x-form.input
+                                                for="excerpt"
+                                                type="text"
+                                                value="{{ old('excerpt', $post->excerpt) }}"
+                                        />
                                     </x-form.label>
                                 </div>
 
-                                {{--Blog Post Head Image--}}
+                                {{-- Blog Post Head Image --}}
                                 <div>
                                     <x-form.label class="mb-2.5">
                                         Post Image
                                     </x-form.label>
 
-                                    {{--File Upload--}}
+                                    {{-- File Upload --}}
                                     <x-form.file-upload :post="$post" />
                                 </div>
 
@@ -65,7 +83,7 @@
                     </x-auth.container>
 
                     <x-auth.container>
-                        {{--Content--}}
+                        {{-- Content --}}
                         <div class="md:grid md:grid-cols-2 md:gap-6">
                             <div class="md:col-span-1">
                                 <h3 class="text-xl font-medium leading-6 text-gray-800 dark:text-gray-400">Content</h3>
@@ -86,25 +104,40 @@
                             </div>
 
                             <div class="space-y-8 mt-5 md:mt-0 md:col-span-2">
-                                {{--Tags--}}
+                                {{-- Tags --}}
                                 <div>
-                                    <x-form.label class="mb-2.5" for="tags"></x-form.label>
-                                    <x-form.multiple-select id="tags" placeholder="Add some tags...">
-                                        @foreach($tags as $tag)
-                                            <option {{ $post->tags->firstWhere('id', $tag->id) ? 'selected' : '' }}
-                                                    value="{{ $tag->id }}">{{ $tag->name }}</option>
+                                    <x-form.label
+                                            class="mb-2.5"
+                                            for="tags"
+                                    ></x-form.label>
+                                    <x-form.multiple-select
+                                            id="tags"
+                                            placeholder="Add some tags..."
+                                    >
+                                        @foreach ($tags as $tag)
+                                            <option
+                                                    value="{{ $tag->id }}"
+                                                    {{ $post->tags->firstWhere('id', $tag->id) ? 'selected' : '' }}
+                                            >{{ $tag->name }}</option>
                                         @endforeach
                                     </x-form.multiple-select>
                                 </div>
 
-                                {{--State--}}
+                                {{-- State --}}
                                 <div>
                                     <div>
-                                        <x-form.label class="mb-2.5" for="is_published" text="State"></x-form.label>
+                                        <x-form.label
+                                                class="mb-2.5"
+                                                for="is_published"
+                                                text="State"
+                                        ></x-form.label>
                                     </div>
                                     <div class="flex items-start">
                                         <div class="flex items-center h-5">
-                                            <x-form.checkbox :checked="$post->is_published === \App\Enums\PostStatus::Published" name="is_published" />
+                                            <x-form.checkbox
+                                                    name="is_published"
+                                                    :checked="$post->is_published === PostStatus::Published"
+                                            />
                                         </div>
                                         <div class="ml-3 text-sm">
                                             <p class="text-gray-500 dark:text-gray-400">Choose if the post will be published, or a draft</p>
@@ -115,7 +148,7 @@
                         </div>
                     </x-auth.container>
 
-                    {{--Buttons--}}
+                    {{-- Buttons --}}
                     <div class="mx-auto w-11/12 sm:w-full flex justify-end space-x-3 sm:px-6 lg:px-8">
                         <x-form.button.cancel>Cancel</x-form.button.cancel>
                         <x-form.button.submit>Submit</x-form.button.submit>
