@@ -1,19 +1,29 @@
-{{--Template--}}
+{{-- Template --}}
 <div>
-    <div x-data x-init="easyMDE">
+    <div
+        x-data
+        x-init="easyMDE"
+    >
         <label for="{{ $id }}"></label>
-        <textarea {{ $attributes }} id="{{ $id }}" name="{{ $name ?? $id }}">{{ old($name ?? $id, $slot) }}</textarea>
+        <textarea
+            id="{{ $id }}"
+            name="{{ $name ?? $id }}"
+            {{ $attributes }}
+        >{{ old($name ?? $id, $slot) }}</textarea>
     </div>
 </div>
-{{--End Template--}}
+{{-- End Template --}}
 
 @push('stylesheets')
-    {{--I know it's not a stylesheet, but placing this before </body> doesn't work--}}
+    {{-- I know it's not a stylesheet, but placing this before </body> doesn't work --}}
     <script src="https://unpkg.com/easymde/dist/easymde.min.js"></script>
 @endpush
 
 @push('scripts')
-    <script async defer>
+    <script
+        async
+        defer
+    >
         const easyMDE = new EasyMDE({
             element: document.getElementById('{{ $id }}'),
             imageUploadFunction: (file, onSuccess, onError) => {
