@@ -1,9 +1,9 @@
 <div>
-    @if($format === 'short')
+    @if ($format === 'short')
         <x-posts-list.short :posts="$posts"></x-posts-list.short>
     @endif
 
-    @if($format === 'full')
+    @if ($format === 'full')
         <x-posts-list.full :posts="$posts"></x-posts-list.full>
     @endif
 </div>
