@@ -1,8 +1,7 @@
 @component('mail::message')
-You've received a new message from **{{ $name }}**
+    You've received a new message from **{{ $name }}**
 
-@component('mail::panel')
-{{ $message }}
-@endcomponent
-
+    @component('mail::panel')
+        {{ $message }}
+    @endcomponent
 @endcomponent
