@@ -1,6 +1,12 @@
-<x-layout.main container subTitle="Chris Mellors' Website and Blog">
+<x-layout.main
+    container
+    subTitle="Chris Mellors' Website and Blog"
+>
     @section('meta-description')
-        <meta name="description" content="The personal website and blog writings of Chris Mellor">
+        <meta
+            name="description"
+            content="The personal website and blog writings of Chris Mellor"
+        >
     @endsection
     <x-header></x-header>
     <x-main></x-main>
